file splitted into utils fct ; pytest added to env ; functional test implemented
