+++ get_data/upload_to_db.py
@@ -1,12 +1,8 @@
 from pathlib import Path
 import json
-import os
-import boto3
-from botocore.exceptions import ClientError
-from elasticsearch import Elasticsearch
-from elasticsearch import helpers
 
-from get_data import cluster_param
+from .utils import s3_utils
+from .utils import es_utils
 
 
 def is_single_label(label):
@@ -16,93 +12,9 @@ def is_single_label(label):
     return True
 
 
-def file_exist_in_bucket(s3, bucket, key):
-    """check if key already exists in bucket. Return True if exists, False otherwise."""
-    try:
-        s3.Object(bucket, key).load()
-    except ClientError as e:
-        return int(e.response['Error']['Code']) != 404
-    return True
-
-
-def upload_label_to_s3(l_label, s3_bucket_name, prefix="", overwrite=False):
-    """
-    Upload picture to s3 bucket.
-    Note that credential to access the s3 bucket is retrieved from env variable (see variable name in the code)
-    :param l_label:             [list]      List of labels. Each label shall contains the following keys:
-                                            "file_name": the name of the picture file
-                                            "location": path to the picture directory
-                                            "img_id": id that will be used to index the picture (shall be unique)
-    :param s3_bucket_name:      [string]    Name of the bucket
-    :param prefix:              [string]    Prefix for every picture
-    :param overwrite:           [bool]      If True, existing pic will be overwritten by new one sharing the same img_id
-    :return:                    [tuple]     list of picture id successfully upladed, list of picture id failed to upload
-    """
-    access_key_id = os.environ["PATATE_S3_KEY_ID"]
-    access_key = os.environ["PATATE_S3_KEY"]
-    already_exist_pic = []
-    upload_success = []
-    s3 = boto3.resource("s3", aws_access_key_id=access_key_id, aws_secret_access_key=access_key)
-    for label in l_label:
-        picture = Path(label["location"]) / label["file_name"]
-        pic_id = label["img_id"]
-        key = prefix + pic_id
-        if not overwrite and file_exist_in_bucket(s3, s3_bucket_name, key):
-            print(f'  --> Can\'t upload file "{picture}" because key "{key}" already exists in bucket "{s3_bucket_name}"')
-            already_exist_pic.append(pic_id)
-        else:
-            s3.meta.client.upload_file(str(picture), s3_bucket_name, key)
-            upload_success.append(pic_id)
-    return upload_success, already_exist_pic
-
-
-def gen_bulk_doc(l_label, index, op_type):
-    """Yield well formatted document for bulk upload to ES"""
-    for label in l_label:
-        yield {
-            "_index": index,
-            "_type": "document",
-            "_op_type": op_type,
-            "_id": label["img_id"],
-            "_source": label
-        }
-
-
-def upload_to_es(l_label, index, host_ip, port, update=False):
-    """
-    Upload all label in l_label to Elasticsearch cluster.
-    Note that credential to access the cluster is retrieved from env variable (see variable name in the code)
-    :param l_label:             [list]      List of labels. Each label shall contains the following keys:
-                                            "file_name": the name of the picture file
-                                            "location": path to the picture directory
-                                            "img_id": id that will be used to index the picture (shall be unique)
-    :param index:               [string]    Name of the index to use for indexing labels
-    :param host_ip:             [string]    Public ip of the Elasticsearch host server
-    :param port:                [int]       Port open for Elasticsearch on host server (typically 9200)
-    :param update:              [bool]      If True, existing document with same img_id will be overwritten by new ones
-    :return:                    [list]      list of failed to upload picture id
-    """
-    try:
-        user = os.environ["ES_USER_ID"]
-        pwd = os.environ["ES_USER_PWD"]
-    except KeyError as err:
-        print("  --> Warning: Elasticsearch 'user' and/or password not found. Trying connection without authentication")
-        user = ""
-        pwd = ""
-    es = Elasticsearch([{"host": host_ip, "port": port}], http_auth=(user, pwd))
-    op_type = "index" if update else "create"
-    success, errors = helpers.bulk(es, gen_bulk_doc(l_label, index, op_type), request_timeout=60, raise_on_error=False)
-    failed_doc_id = []
-    for error in errors:
-        for error_type in error:
-            failed_doc_id.append(error[error_type]["_id"])
-            print(f'  --> Couldn\'t upload document {failed_doc_id[-1]} because:{error[error_type]["error"]["reason"]}')
-    return failed_doc_id
-
-
 def filter_out_missing_pic(label_list):
     """
-    Check if all pictures from the label_list exists.
+    Check if all pictures referenced in the label_list exists.
     Return a list of valid_label (picture exists) and a list of missing picture id
     :param label_list:      [list]  list of label
     :return:                [tuple] list of valid label (picture exists) and list of missing pictures id
@@ -123,11 +35,13 @@ def edit_label(l_label, field, value):
         label[field] = value
 
 
-def print_synthesis(upload_bucket_path, missing_pic, already_exist_pic, s3_success, failed_es):
+def upload_synthesis(upload_bucket_path, missing_pic, already_exist_pic, s3_success, failed_es):
+    success = len(s3_success) - len(failed_es)
+    fail = len(missing_pic) + len(already_exist_pic) + len(failed_es)
     print('----------------------------------------')
     print('Upload Completed !')
-    print(f'{len(s3_success) - len(failed_es)} picture(s) were successful uploaded.')
-    print(f'{len(missing_pic) + len(already_exist_pic) + len(failed_es)} upload failed.')
+    print(f'{success} picture(s) were successful uploaded.')
+    print(f'{fail} upload failed.')
     print(f'Upload bucket: "{upload_bucket_path}"')
     print(f'Number of picture uploaded to s3 bucket : {len(s3_success)}')
     print(f'Number of failed upload to s3 bucket: {len(missing_pic) + len(already_exist_pic)}')
@@ -138,25 +52,19 @@ def print_synthesis(upload_bucket_path, missing_pic, already_exist_pic, s3_succe
     print(f'Number of failed upload to ES cluster: {len(failed_es)}')
     if len(failed_es) > 0:
         print(f'  --> List of failed pic id: {already_exist_pic}')
+    return success, fail
 
 
-def get_s3_formatted_bucket_path(bucket_name, prefix):
-    """
-    Creates a cleaned and well formatted name and path for upload to s3 bucket from the bucket name and key prefix.
-    :param bucket_name:     [string]        s3 Bucket name
-    :param prefix:          [string]        key prefix (s3 sub-folder)
-    :return:                [tuple of str]  clean_full_path, clean_buck_name, clean_key_prefix
-
-    For example:
-    >>> get_s3_formatted_bucket_path("my-bucket/", "/sub/bucket//directory/with/typo")
-    ("my-bucket/sub/bucket/directory/with/typo/", "my-bucket", "sub/bucket/directory/with/typo/")
-    """
-    raw_dir = bucket_name + "/" + prefix
-    dir_list_clean = [elm for elm in raw_dir.split("/") if bool(elm)]
-    clean_full_path = "/".join(dir_list_clean) + "/"
-    clean_key_prefix = "/".join(dir_list_clean[1:]) + "/" if len(dir_list_clean) > 1 else ""
-    clean_bucket_name = dir_list_clean[0]
-    return clean_full_path, clean_bucket_name, clean_key_prefix
+def get_label_list_from_file(file):
+    """Open and read file containing the label(s). Return a list of label or None on errors."""
+    if not Path(file).is_file():
+        print(f'File "{file}" can\'t be found.')
+        return None
+    with Path(file).open(mode='r', encoding='utf-8') as fp:
+        l_label = json.load(fp)
+    if is_single_label(l_label):
+        l_label = [l_label]
+    return l_label
 
 
 def upload_to_db(label_file, bucket_name, key_prefix, es_host_ip, es_port, es_index, overwrite=False):
@@ -184,32 +92,21 @@ def upload_to_db(label_file, bucket_name, key_prefix, es_host_ip, es_port, es_in
     :param es_index:            [string]    Name of the index to use
     :param overwrite:           [bool]      If True will overwrite picture is s3 and label in ES if same img_id is found
                                             If False (default), only non existing img_id picture will be uploaded
+    :return:                    [tuple]     (int) Number of successful upload, (int) number of failed upload
+                                            Return None on error
     """
-    if not Path(label_file).is_file():
-        print(f'File "{label_file}" can\'t be found.')
-        exit(0)
-    with Path(label_file).open(mode='r', encoding='utf-8') as fp:
-        l_label = json.load(fp)
-    if is_single_label(l_label):
-        l_label = [l_label]
+    l_label = get_label_list_from_file(label_file)
+    if l_label is None:
+        return None, None
     print(f'Looking for pictures...')
     l_label, missing_pic = filter_out_missing_pic(l_label)
     print(f'Uploading to s3...')
-    upload_bucket_dir, bucket_name, key_prefix = get_s3_formatted_bucket_path(bucket_name, key_prefix)
-    s3_upload_success, already_exist_pic = upload_label_to_s3(l_label, bucket_name, key_prefix, overwrite)
+    upload_bucket_dir, bucket_name, key_prefix = s3_utils.get_s3_formatted_bucket_path(bucket_name, key_prefix)
+    s3_upload_success, already_exist_pic = s3_utils.upload_to_s3_from_label(l_label, bucket_name, key_prefix, overwrite)
     l_label = [label for label in l_label if label["img_id"] in s3_upload_success]
     edit_label(l_label, "location", upload_bucket_dir)
     print(f'Uploading to Elasticsearch cluster...')
-    failed_es_upload = upload_to_es(l_label=l_label, index=es_index, host_ip=es_host_ip, port=es_port, update=overwrite)
-    print_synthesis(upload_bucket_dir, missing_pic, already_exist_pic, s3_upload_success, failed_es_upload)
-
-
-if __name__ == "__main__":
-    label_file = "sample/labels.json"
-    # label_file = "sample/single_label.json"
-    s3_bucket = "032854191254-patate"
-    key_prefix = ""
-    es_ip_host = cluster_param.ES_HOST_IP
-    es_port_host = cluster_param.ES_HOST_PORT
-    es_index_name = cluster_param.ES_INDEX
-    upload_to_db(label_file, s3_bucket, key_prefix, es_ip_host, es_port_host, es_index_name, overwrite=True)
+    failed_es_upload = es_utils.upload_to_es(
+        l_label=l_label, index=es_index, host_ip=es_host_ip, port=es_port, update=overwrite)
+    success, fail = upload_synthesis(upload_bucket_dir, missing_pic, already_exist_pic, s3_upload_success, failed_es_upload)
+    return success, fail
